ddgfadssdefd

+++ src/main/java/L_Ender/cataclysm/entity/Ignis_Entity.java
@@ -1398,7 +1398,8 @@ public class Ignis_Entity extends Boss_monster {
                         AABB aabb = this.getBoundingBox().inflate(0.2D);
                         for (BlockPos blockpos : BlockPos.betweenClosed(Mth.floor(aabb.minX), Mth.floor(this.getY()), Mth.floor(aabb.minZ), Mth.floor(aabb.maxX), Mth.floor(aabb.maxY), Mth.floor(aabb.maxZ))) {
                             BlockState blockstate = this.level.getBlockState(blockpos);
-                            if (blockstate.canEntityDestroy(this.level, blockpos, this) && !blockstate.is(ModTag.IGNIS_IMMUNE) && net.minecraftforge.event.ForgeEventFactory.onEntityDestroyBlock(this, blockpos, blockstate)) {
+                            FluidState fluidState = level.getFluidState(blockpos);
+                            if (blockstate.getMaterial() != Material.AIR && blockstate.canEntityDestroy(this.level, blockpos, this) && fluidState.isEmpty() && !blockstate.is(ModTag.IGNIS_IMMUNE) && net.minecraftforge.event.ForgeEventFactory.onEntityDestroyBlock(this, blockpos, blockstate)) {
                                 flag = this.level.destroyBlock(blockpos, true, this) || flag;
                             }
                         }
